Create reg salt util
Use the reg salt util to get grains
Use the reg salt util in the reg module
Use the reg salt util in the reg state
Use the reg salt util to get the PowerShell version
Detect version of PowerShell instead of version of Windows to determine
ConvertTo-JSON usage
